Address review feedback (round 2): test hygiene + JSDoc placement
Apply review nits from PR #515 round-2 reviewer pass:

- **JSDoc misattachment** (`git.ts`): the long "label must be
  non-empty / unique / filesystem-safe" block was positioned above
  `DETACHED_BRANCH_PREFIX` instead of `detachedShaLabel` — IDE
  tooling attaches a JSDoc comment to the immediately-following
  declaration, so the explanation was attached to the wrong symbol.
  Swap the order: short prefix-doc on the constant, long invariants-doc
  on the function.
- **Un-export `detachedShaLabel`**: the only caller outside `git.ts`
  was the integration test (round-1 nit response). Round-2 reviewer
  flagged that as an internal-module import in a black-box test.
  Reverting the export keeps the production API surface smaller; the
  test now inlines `"detached-" + olderSha.slice(0, 7)` with a comment
  explaining the deliberate redundancy (format also pinned by
  `git.test.ts`, so silent drift would still fail loudly).
  `DETACHED_BRANCH_PREFIX` stays exported — it has a real non-test
  caller in `workspaces.remove`.
- **Extract server-boot helpers**: created
  `apps/web/tests/helpers/server.ts` with the canonical versions of
  `createTmpHome`, `getRandomPort`, and `startServer` (including the
  SIGKILL fallback that had only existed in `task-cleanup.test.ts`).
  The new test uses the shared helper; the three pre-existing tests
  that inline copies (`trpc.test.ts`, `task-cleanup.test.ts`,
  `trpc-batch-url.test.ts`) are intentionally not migrated in this
  PR to keep the diff focused on the bug fix — follow-up tracked.

No behaviour change. 21/21 tests pass, lint clean.
