+++ apps/web/src/lib/git.ts
@@ -180,6 +180,14 @@ export async function listWorktrees(repoPath: string): Promise<WorktreeInfo[]> {
 }
 
 /**
+ * Prefix used by `detachedShaLabel`. Exported so callers that need to
+ * recognise the synthetic label (e.g. `workspaces.remove` skipping
+ * `git branch -D` for a non-real ref) can do so without re-spelling
+ * the literal `"detached-"` in two places.
+ */
+export const DETACHED_BRANCH_PREFIX = "detached-";
+
+/**
  * Build a fallback "branch" label for a detached HEAD that isn't mid-rebase
  * (e.g. a checked-out tag, a raw commit SHA, or a post-rebase/post-bisect
  * state). The label must be:
@@ -206,15 +214,7 @@ export async function listWorktrees(repoPath: string): Promise<WorktreeInfo[]> {
  * would still collide, but git doesn't actually produce empty `HEAD` lines
  * in practice — the marker is defensive, not load-bearing.
  */
-/**
- * Prefix used by `detachedShaLabel`. Exported so callers that need to
- * recognise the synthetic label (e.g. `workspaces.remove` skipping
- * `git branch -D` for a non-real ref) can do so without re-spelling
- * the literal `"detached-"` in two places.
- */
-export const DETACHED_BRANCH_PREFIX = "detached-";
-
-export function detachedShaLabel(head: string): string {
+function detachedShaLabel(head: string): string {
   const sha = head.trim().slice(0, 7);
   return sha ? `${DETACHED_BRANCH_PREFIX}${sha}` : "detached";
 }
